+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.12.0] - 2024-12-14
 
 ### Added
-    * Support for local/domain prekey usage for decryption
+    * Support for local/domain credkey usage for masterkey decryption
     * Slack support to statekeys/cookies commands (@Lee Christensen)
     * Ability  to specify local state file (@Lee Christensen)
     * RPC (MS-BKUP) masterkey retrieval (@Kiblyn11)